fixed the long decimal string. inputs are now limited to 2 decimal places

+++ wallet/forms.py
@@ -3,7 +3,7 @@ from wallet.models import Wallet, currencies
 
 
 class TransferForm(forms.Form):
-	amount=forms.FloatField(min_value=0,label='',
+	amount=forms.DecimalField(min_value=0,decimal_places=2,label='',
 		widget=forms.NumberInput(attrs={'placeholder':'Transfer Amount'}))
 	currency=forms.ChoiceField(currencies,label='')
 	user=forms.IntegerField(label='',
